+++ docs/specs/static.yml
@@ -27,7 +27,7 @@ inputs:
 outputs: 
   docs/a/index.html:
   .errors.log: |
-    {"message_severity":"warning","code":"liquid-not-found"}
+    {"message_severity":"warning","code":"liquid-not-found","file":"docs/a.md","line":1,"column":1}
 ---
 # Source file name ending with spaces should not crash
 inputs:
